requirements.txt
db-sqlite3>=0.0.1
discord.py>=2.1.0